[backend] refactor(refreshConnectivity): move refresh connectivity to a new app
