+++ apps/portal-api/src/modules/registration/registration.connectivity.app.ts
@@ -1,0 +1,202 @@
+import {
+  PlatformIdentifier,
+  PlatformRegistrationConnectivityStatus,
+  RefreshPlatformRegistrationConnectivityStatusAllTenantsInput,
+  RefreshPlatformRegistrationConnectivityStatusInput,
+  RefreshPlatformRegistrationConnectivityStatusSingleTenantInput,
+  ServiceConfigurationStatus,
+  TenantStatus,
+} from '../../__generated__/resolvers-types';
+import { logApp } from '../../utils/app-logger.util';
+import { BadRequestErrorCode, ErrorCode } from '../../utils/error/error.code';
+import { RequiredPlatformVersions } from '../../utils/required-platform-version';
+import { doesVersionSatisfy, isValidVersion } from '../../utils/versioning';
+import { PlatformConfiguration } from './registration.domain';
+import { isTenantIdRequired } from './registration.helper';
+import { ServiceConfigurationDomain } from './service-configuration/service-configuration.domain';
+
+const refreshConnectivityStatus = async ({
+  platform_id,
+  token,
+  platform_version,
+  url,
+  tenant_id,
+  platform_identifier,
+}: {
+  platform_id: string;
+  token: string;
+  platform_version: string;
+  url?: string;
+  tenant_id?: string;
+  platform_identifier?: PlatformIdentifier;
+}): Promise<{ status: PlatformRegistrationConnectivityStatus }> => {
+  if (!isValidVersion(platform_version)) {
+    throw new Error(ErrorCode.InvalidPlatformVersion);
+  }
+
+  if (
+    !tenant_id &&
+    platform_identifier &&
+    isTenantIdRequired(platform_identifier, platform_version)
+  ) {
+    throw new Error(BadRequestErrorCode.TenantIdMandatory);
+  }
+
+  let serviceConfiguration =
+    await ServiceConfigurationDomain.loadConfigurationByPlatformAndToken({
+      platform_id: platform_id,
+      token,
+      tenant_id,
+    });
+
+  // manage upgrades from older platform without tenant to newer platform with tenant
+  if (!serviceConfiguration && tenant_id && platform_identifier) {
+    const configWithoutTenant =
+      await ServiceConfigurationDomain.loadConfigurationByPlatformAndToken({
+        platform_id: platform_id,
+        token,
+        withoutTenantId: true,
+      });
+
+    const existingConfig = configWithoutTenant?.config as
+      | PlatformConfiguration
+      | undefined;
+
+    if (
+      configWithoutTenant &&
+      !isTenantIdRequired(platform_identifier, existingConfig?.platform_version)
+    ) {
+      const updatedConfig = { ...existingConfig, tenant_id };
+      await ServiceConfigurationDomain.updateConfiguration(
+        configWithoutTenant.service_instance_id,
+        { config: updatedConfig }
+      );
+      serviceConfiguration = { ...configWithoutTenant, config: updatedConfig };
+    }
+  }
+
+  if (!serviceConfiguration) {
+    if (!platform_identifier) {
+      return { status: PlatformRegistrationConnectivityStatus.Inactive };
+    }
+
+    const requiredVersionForNotFoundStatus =
+      RequiredPlatformVersions.RefreshConnectivityStatusSendsNotFound[
+        platform_identifier
+      ];
+
+    const shouldSendNotFoundStatus = doesVersionSatisfy({
+      givenVersion: platform_version,
+      requiredVersion: requiredVersionForNotFoundStatus,
+    });
+
+    return {
+      status: shouldSendNotFoundStatus
+        ? PlatformRegistrationConnectivityStatus.NotFound
+        : PlatformRegistrationConnectivityStatus.Inactive,
+    };
+  }
+
+  const existingConfig = serviceConfiguration.config;
+  const hasConfigChanged =
+    existingConfig['platform_version'] !== platform_version ||
+    (url && existingConfig['url'] !== url);
+  if (hasConfigChanged) {
+    await ServiceConfigurationDomain.updateConfiguration(
+      serviceConfiguration.service_instance_id,
+      {
+        config: {
+          ...(existingConfig as object),
+          platform_version,
+          ...(url ? { url } : {}),
+        },
+      }
+    );
+  }
+
+  return {
+    status:
+      serviceConfiguration.status === ServiceConfigurationStatus.Active
+        ? PlatformRegistrationConnectivityStatus.Active
+        : PlatformRegistrationConnectivityStatus.Inactive,
+  };
+};
+
+export const registrationConnectivityApp = {
+  refreshPlatformRegistrationConnectivityStatus: async (
+    input: RefreshPlatformRegistrationConnectivityStatusInput
+  ): Promise<{ status: PlatformRegistrationConnectivityStatus }> => {
+    return refreshConnectivityStatus({
+      platform_identifier: input.platformIdentifier,
+      platform_version: input.platformVersion,
+      platform_id: input.platformId,
+      token: input.token,
+    });
+  },
+
+  refreshPlatformRegistrationConnectivityStatusSingleTenant: async (
+    input: RefreshPlatformRegistrationConnectivityStatusSingleTenantInput
+  ): Promise<{ status: PlatformRegistrationConnectivityStatus }> => {
+    return refreshConnectivityStatus({
+      platform_identifier: input.platformIdentifier,
+      platform_version: input.platformVersion,
+      platform_id: input.platformId,
+      token: input.token,
+      url: input.url,
+      tenant_id: input.tenantId,
+    });
+  },
+
+  refreshPlatformRegistrationConnectivityStatusAllTenants: async (
+    input: RefreshPlatformRegistrationConnectivityStatusAllTenantsInput
+  ): Promise<{ statuses: TenantStatus[] }> => {
+    const results = await Promise.allSettled(
+      input.tenants.map((tenant) =>
+        refreshConnectivityStatus({
+          platform_identifier: input.platformIdentifier,
+          platform_version: input.platformVersion,
+          platform_id: input.platformId,
+          token: tenant.token,
+          url: tenant.url,
+          tenant_id: tenant.tenantId,
+        }).then(({ status }) => ({ tenantId: tenant.tenantId, status }))
+      )
+    );
+    const statuses = results.map((result, index) => {
+      if (result.status === 'fulfilled') {
+        return result.value;
+      }
+      logApp.error(
+        `Failed to refresh connectivity status for tenant ${input.tenants[index].tenantId}`,
+        { error: result.reason }
+      );
+      return {
+        tenantId: input.tenants[index].tenantId,
+        status: PlatformRegistrationConnectivityStatus.Inactive,
+      };
+    });
+
+    const knownTenantIds = input.tenants.map((t) => t.tenantId);
+    const staleConfigurations =
+      await ServiceConfigurationDomain.loadActiveConfigurationsByPlatformExcludingTenants(
+        input.platformId,
+        knownTenantIds
+      );
+
+    await Promise.all(
+      staleConfigurations.map((config) =>
+        ServiceConfigurationDomain.updateConfiguration(
+          config.service_instance_id,
+          { status: ServiceConfigurationStatus.Inactive }
+        ).catch((error) => {
+          logApp.error(
+            `Failed to deactivate stale configuration for service instance ${config.service_instance_id}`,
+            { error }
+          );
+        })
+      )
+    );
+
+    return { statuses };
+  },
+};
